+++ src/config/configUtils.test.ts
@@ -1,6 +1,7 @@
 import { expect } from "chai";
 import { getMatchingPluginConfig, MatchParams, mergeConfig } from "./configUtils";
 import { PluginOptions } from "../index";
+import { BasePluginType } from "../plugins/pluginTypes";
 
 describe("configUtils", () => {
   describe("mergeConfig", () => {
@@ -55,12 +56,14 @@ describe("configUtils", () => {
   });
 
   describe("getMatchingPluginConfig", () => {
-    interface ISharedPluginConfig {
-      value: number;
-      hasAccess: boolean;
+    interface SharedPluginType extends BasePluginType {
+      config: {
+        value: number;
+        hasAccess: boolean;
+      };
     }
 
-    const sharedPluginOptions: PluginOptions<ISharedPluginConfig> = {
+    const sharedPluginOptions: PluginOptions<SharedPluginType> = {
       config: {
         value: 5,
         hasAccess: false,
@@ -119,30 +122,30 @@ describe("configUtils", () => {
     };
 
     it("should use defaults with empty match params", () => {
-      const matchedConfig = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {});
+      const matchedConfig = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {});
       expect(matchedConfig.value).to.equal(5);
       expect(matchedConfig.hasAccess).to.equal(false);
     });
 
     it("should match levels", () => {
-      const matchedConfig = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         level: 60,
       });
       expect(matchedConfig.hasAccess).to.equal(true);
     });
 
     it("should require all level conditions to apply", () => {
-      const matchedConfig = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         level: 35,
       });
       expect(matchedConfig.hasAccess).to.equal(false);
     });
 
     it("should match channels and accept any specified channel", () => {
-      const matchedConfig1 = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig1 = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         channelId: "1100",
       });
-      const matchedConfig2 = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig2 = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         channelId: "1200",
       });
       expect(matchedConfig1.value).to.equal(10);
@@ -150,10 +153,10 @@ describe("configUtils", () => {
     });
 
     it("should match categories and accept any specified category", () => {
-      const matchedConfig1 = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig1 = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         categoryId: "9100",
       });
-      const matchedConfig2 = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig2 = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         categoryId: "9200",
       });
       expect(matchedConfig1.value).to.equal(120);
@@ -161,17 +164,17 @@ describe("configUtils", () => {
     });
 
     it("should match users", () => {
-      const matchedConfig = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         userId: "2100",
       });
       expect(matchedConfig.value).to.equal(15);
     });
 
     it("should match roles", () => {
-      const matchedConfig1 = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig1 = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         memberRoles: ["3100"],
       });
-      const matchedConfig2 = getMatchingPluginConfig<ISharedPluginConfig>(null as any, sharedPluginOptions, {
+      const matchedConfig2 = getMatchingPluginConfig<SharedPluginType>(null as any, sharedPluginOptions, {
         memberRoles: ["3100", "3200"],
       });
       expect(matchedConfig1.value).to.equal(20); // has 3100, and no 3200 -> match
@@ -179,16 +182,17 @@ describe("configUtils", () => {
     });
 
     it("custom resolver", () => {
-      interface ICustomPluginConfig {
-        value: number;
+      interface CustomPluginType extends BasePluginType {
+        config: {
+          value: number;
+        };
+        customOverrideCriteria: {
+          targetUserId?: string;
+          targetChannelId?: string;
+        };
       }
 
-      interface ICustomOverrideCriteria {
-        targetUserId?: string;
-        targetChannelId?: string;
-      }
-
-      const customPluginOptions: PluginOptions<ICustomPluginConfig, ICustomOverrideCriteria> = {
+      const customPluginOptions: PluginOptions<CustomPluginType> = {
         config: {
           value: 5,
         },
@@ -221,13 +225,17 @@ describe("configUtils", () => {
         ],
       };
 
-      const customResolver = (pluginData, criteria: ICustomOverrideCriteria, matchParams: MatchParams): boolean => {
+      const customResolver = (
+        pluginData,
+        criteria: CustomPluginType["customOverrideCriteria"],
+        matchParams: MatchParams
+      ): boolean => {
         if (criteria.targetUserId && matchParams.extra.targetUserId !== criteria.targetUserId) return false;
         if (criteria.targetChannelId && matchParams.extra.targetChannelId !== criteria.targetChannelId) return false;
         return true;
       };
 
-      const matchedConfig1 = getMatchingPluginConfig<ICustomPluginConfig>(
+      const matchedConfig1 = getMatchingPluginConfig<CustomPluginType>(
         null as any,
         customPluginOptions,
         {
@@ -237,7 +245,7 @@ describe("configUtils", () => {
         },
         customResolver
       );
-      const matchedConfig2 = getMatchingPluginConfig<ICustomPluginConfig>(
+      const matchedConfig2 = getMatchingPluginConfig<CustomPluginType>(
         null as any,
         customPluginOptions,
         {
@@ -247,7 +255,7 @@ describe("configUtils", () => {
         },
         customResolver
       );
-      const matchedConfig3 = getMatchingPluginConfig<ICustomPluginConfig>(
+      const matchedConfig3 = getMatchingPluginConfig<CustomPluginType>(
         null as any,
         customPluginOptions,
         {
@@ -257,7 +265,7 @@ describe("configUtils", () => {
         },
         customResolver
       );
-      const matchedConfig4 = getMatchingPluginConfig<ICustomPluginConfig>(
+      const matchedConfig4 = getMatchingPluginConfig<CustomPluginType>(
         null as any,
         customPluginOptions,
         {
@@ -360,10 +368,13 @@ describe("configUtils", () => {
     });
 
     it("'all' special criterion", () => {
-      interface IThisOptions {
-        value: number;
+      interface PluginType extends BasePluginType {
+        config: {
+          value: number;
+        };
       }
-      const pluginOpts: PluginOptions<IThisOptions> = {
+
+      const pluginOpts: PluginOptions<PluginType> = {
         config: {
           value: 5,
         },
@@ -385,15 +396,15 @@ describe("configUtils", () => {
         ],
       };
 
-      const matchedConfig1 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts, {
+      const matchedConfig1 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts, {
         userId: "1000",
         level: 75,
       });
-      const matchedConfig2 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts, {
+      const matchedConfig2 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts, {
         userId: "1000",
         level: 120,
       });
-      const matchedConfig3 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts, {
+      const matchedConfig3 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts, {
         userId: "1000",
         level: 25,
       });
@@ -404,10 +415,13 @@ describe("configUtils", () => {
     });
 
     it("'any' special criterion", () => {
-      interface IThisOptions {
-        value: number;
+      interface PluginType extends BasePluginType {
+        config: {
+          value: number;
+        };
       }
-      const pluginOpts: PluginOptions<IThisOptions> = {
+
+      const pluginOpts: PluginOptions<PluginType> = {
         config: {
           value: 5,
         },
@@ -428,13 +442,13 @@ describe("configUtils", () => {
         ],
       };
 
-      const matchedConfig1 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts, {
+      const matchedConfig1 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts, {
         level: 15,
       });
-      const matchedConfig2 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts, {
+      const matchedConfig2 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts, {
         level: 95,
       });
-      const matchedConfig3 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts, {
+      const matchedConfig3 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts, {
         level: 50,
       });
 
@@ -444,11 +458,13 @@ describe("configUtils", () => {
     });
 
     it("'not' special criterion", () => {
-      interface IThisOptions {
-        value: number;
+      interface PluginType extends BasePluginType {
+        config: {
+          value: number;
+        };
       }
 
-      const pluginOpts1: PluginOptions<IThisOptions> = {
+      const pluginOpts1: PluginOptions<PluginType> = {
         config: {
           value: 5,
         },
@@ -465,7 +481,7 @@ describe("configUtils", () => {
         ],
       };
 
-      const pluginOpts2: PluginOptions<IThisOptions> = {
+      const pluginOpts2: PluginOptions<PluginType> = {
         config: {
           value: 5,
         },
@@ -489,7 +505,7 @@ describe("configUtils", () => {
         ],
       };
 
-      const pluginOpts3: PluginOptions<IThisOptions> = {
+      const pluginOpts3: PluginOptions<PluginType> = {
         config: {
           value: 5,
         },
@@ -506,21 +522,21 @@ describe("configUtils", () => {
         ],
       };
 
-      const matchedConfig1 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts1, {
+      const matchedConfig1 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts1, {
         userId: "1234",
       });
-      const matchedConfig2 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts1, {
+      const matchedConfig2 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts1, {
         userId: "5678",
       });
 
       expect(matchedConfig1.value).to.equal(5);
       expect(matchedConfig2.value).to.equal(10);
 
-      const matchedConfig3 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts2, {
+      const matchedConfig3 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts2, {
         level: 95,
         userId: "1234",
       });
-      const matchedConfig4 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts2, {
+      const matchedConfig4 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts2, {
         level: 95,
         userId: "5678",
       });
@@ -528,13 +544,13 @@ describe("configUtils", () => {
       expect(matchedConfig3.value).to.equal(5);
       expect(matchedConfig4.value).to.equal(20);
 
-      const matchedConfig5 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts3, {
+      const matchedConfig5 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts3, {
         level: 49,
       });
-      const matchedConfig6 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts3, {
+      const matchedConfig6 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts3, {
         level: 50,
       });
-      const matchedConfig7 = getMatchingPluginConfig<IThisOptions>(null as any, pluginOpts3, {
+      const matchedConfig7 = getMatchingPluginConfig<PluginType>(null as any, pluginOpts3, {
         level: 51,
       });