+++ src/events/eventUtils.ts
@@ -1,12 +1,15 @@
 import { Channel, Guild, GuildChannel, Member, Message, User } from "eris";
 import { KnownEventArguments } from "./eventArguments";
 import { EventListenerBlueprint } from "./EventListenerBlueprint";
+import { BasePluginType } from "../plugins/pluginTypes";
 
 /**
  * An identity function that helps with type hinting.
  * Takes an event listener blueprint as an argument and returns that same blueprint.
  */
-export function asEventListener<T extends string>(blueprint: EventListenerBlueprint<T>): EventListenerBlueprint<T> {
+export function asEventListener<TPluginType extends BasePluginType, TEventName extends string>(
+  blueprint: EventListenerBlueprint<TPluginType, TEventName>
+): EventListenerBlueprint<TPluginType, TEventName> {
   return blueprint;
 }