+++ src/plugins/pluginTypes.ts
@@ -1,0 +1,6 @@
+import { BasePluginConfig } from "..";
+
+export interface BasePluginType {
+  config: BasePluginConfig;
+  customOverrideCriteria: unknown;
+}